refactor: address PR review feedback
- Move getChangedLineNumbers to diff-utils.ts helper file
- Extract file change tracking to useFileChanges hook
- Move inline style to CSS module (output-view-wrapper)
- Use finally for setIsCompiling(false) in doCompile
- Pass changedLineNumbers through FileOutputViewerProps instead of
  cloning RawFileViewer with a modified render function
