+++ packages/playground/src/react/output-view/file-viewer.tsx
@@ -1,12 +1,12 @@
 import { FolderListRegular } from "@fluentui/react-icons";
 import { Pane, SplitPane } from "@typespec/react-components";
-import { useCallback, useEffect, useMemo, useRef, useState } from "react";
+import { useCallback, useEffect, useMemo, useState } from "react";
 import { FileBreadcrumb } from "../breadcrumb/index.js";
 import { FileOutput } from "../file-output/file-output.js";
 import { FileTreeExplorer } from "../file-tree/index.js";
 import { OutputTabs } from "../output-tabs/output-tabs.js";
 import type { FileOutputViewer, OutputViewerProps, ProgramViewer } from "../types.js";
-import { getChangedLineNumbers } from "../typespec-editor.js";
+import { useFileChanges } from "./use-file-changes.js";
 
 import style from "./output-view.module.css";
 
@@ -21,9 +21,7 @@ const FileViewerComponent = ({
 }) => {
   const [filename, setFilename] = useState<string>("");
   const [content, setContent] = useState<string>("");
-  const [changedFiles, setChangedFiles] = useState<Set<string>>(new Set());
-  const [changedLines, setChangedLines] = useState<Map<string, number[]>>(new Map());
-  const prevContentsRef = useRef<Map<string, string>>(new Map());
+  const { changedFiles, changedLines } = useFileChanges(program, outputFiles, highlightChanges);
 
   const showFileTree = useMemo(
     () => outputFiles.some((f) => f.includes("/")) || outputFiles.length >= 3,
@@ -37,66 +35,6 @@ const FileViewerComponent = ({
     },
     [program.host],
   );
-
-  // When output files change, diff all file contents against cached versions
-  useEffect(() => {
-    if (!highlightChanges) return;
-    let cancelled = false;
-    async function diffFiles() {
-      const changed = new Set<string>();
-      const lines = new Map<string, number[]>();
-      const newContents = new Map<string, string>();
-
-      // If no files from the new output exist in the cache, this is an emitter
-      // switch or initial load — populate the cache without highlighting.
-      const isEmitterSwitch =
-        prevContentsRef.current.size > 0 &&
-        !outputFiles.some((f) => prevContentsRef.current.has(f));
-
-      let hasAnyChange = false;
-      for (const file of outputFiles) {
-        try {
-          const contents = await program.host.readFile("./tsp-output/" + file);
-          newContents.set(file, contents.text);
-          if (!isEmitterSwitch) {
-            const prev = prevContentsRef.current.get(file);
-            if (prev === undefined && prevContentsRef.current.size > 0) {
-              changed.add(file);
-              // New file: highlight all lines
-              const lineCount = contents.text.split("\n").length;
-              lines.set(
-                file,
-                Array.from({ length: lineCount }, (_, i) => i + 1),
-              );
-              hasAnyChange = true;
-            } else if (prev !== undefined && prev !== contents.text) {
-              changed.add(file);
-              lines.set(file, getChangedLineNumbers(prev, contents.text));
-              hasAnyChange = true;
-            } else if (prev === undefined) {
-              hasAnyChange = true;
-            }
-          } else {
-            hasAnyChange = true;
-          }
-        } catch {
-          // file may not be readable
-        }
-      }
-      if (cancelled) return;
-      // Only update cache and changed state when something actually changed.
-      // This prevents spurious effect re-runs from clearing the highlights.
-      if (hasAnyChange || prevContentsRef.current.size === 0) {
-        prevContentsRef.current = newContents;
-        setChangedFiles(changed);
-        setChangedLines(lines);
-      }
-    }
-    void diffFiles();
-    return () => {
-      cancelled = true;
-    };
-  }, [program, outputFiles, highlightChanges]);
 
   useEffect(() => {
     if (outputFiles.length > 0) {
