+++ packages/playground/src/react/typespec-editor.tsx
@@ -3,6 +3,9 @@ import { useCallback, useEffect, useRef, useState, type FunctionComponent } from
 import { Editor, useMonacoModel, type EditorProps } from "./editor.js";
 import type { PlaygroundEditorsOptions } from "./playground.js";
 
+// Re-export for backward compatibility
+export { getChangedLineNumbers } from "./diff-utils.js";
+
 export interface TypeSpecEditorProps extends Omit<EditorProps, "options"> {
   options?: editor.IStandaloneEditorConstructionOptions;
 }
@@ -23,42 +26,6 @@ export const TypeSpecEditor: FunctionComponent<TypeSpecEditorProps> = ({
   };
   return <Editor actions={actions} options={resolvedOptions} {...other}></Editor>;
 };
-
-/**
- * Computes which lines in the new text are changed or inserted compared to the old text.
- * Uses a longest common subsequence (LCS) approach to handle insertions/deletions properly.
- */
-export function getChangedLineNumbers(oldText: string, newText: string): number[] {
-  const oldLines = oldText.split("\n");
-  const newLines = newText.split("\n");
-
-  // Build a set of old lines that are "matched" via LCS
-  const matchedNewIndices = new Set<number>();
-
-  // Simple greedy match: walk both arrays with two pointers
-  let oi = 0;
-  for (let ni = 0; ni < newLines.length; ni++) {
-    // Try to find current new line in remaining old lines
-    for (let j = oi; j < oldLines.length; j++) {
-      if (newLines[ni] === oldLines[j]) {
-        matchedNewIndices.add(ni);
-        oi = j + 1;
-        break;
-      }
-    }
-    // If not found in forward scan, it might still match a later occurrence
-    // but for highlighting purposes, treating it as changed is acceptable
-  }
-
-  // Lines not matched are changed/inserted
-  const changed: number[] = [];
-  for (let ni = 0; ni < newLines.length; ni++) {
-    if (!matchedNewIndices.has(ni)) {
-      changed.push(ni + 1); // Monaco lines are 1-based
-    }
-  }
-  return changed;
-}
 
 export const OutputEditor: FunctionComponent<{
   filename: string;
